Downloads/Deep-Researcher-Agent-main/Deep-Researcher-Agent-main/requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0

# AI and ML Libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
accelerate>=0.20.0

# Vector Database
faiss-cpu>=1.7.4

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Text Processing
nltk>=3.8.0
spacy>=3.6.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0